Centralize nm_otel test utilities
Copilot-Session: 546e2593-f484-46cd-bc7c-81cb39dcff99
